First trial to add more extracted info
I tried to add the date and country to the search (it took the info from the last result for all rows, probably because I placed my code outside the loop?). I also did not know how the IPC could be referenced to be also included.
agripatents2 is the document with the extracts
Not happy with results yet.